[EDIT]#252 에러처리
alert 수정

+++ frontend/frontproject/src/pages/storyDetailPage.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
 import { getStory, deleteStory } from "../api/storyApi";
@@ -7,6 +7,7 @@ import BGMPlayer from "../components/storyResult/bgm";
 import AudioPlayer from "../components/storyResult/audio";
 import { voiceAtom, genreAtom } from "../atoms";
 import { useRecoilState, useSetRecoilState } from "recoil";
+import Swal from "sweetalert2";
 
 export default function StoryDetailPage() {
 	// 로그인된 유저
@@ -14,6 +15,7 @@ export default function StoryDetailPage() {
 	const id: number = Number(params.id);
 
 	const navigate = useNavigate();
+
 	const setVoice = useSetRecoilState(voiceAtom);
 	const [genre, setGenre] = useRecoilState(genreAtom);
 	const [storyInfo, setStoryInfo] = useState({
@@ -25,13 +27,9 @@ export default function StoryDetailPage() {
 		voice: "",
 	});
 
-	useEffect(() => {
-		getStoryItem();
-	}, []);
-
 	const [lang, setLang] = useState(true);
 	// 이야기 상세 불러오기
-	const getStoryItem = async () => {
+	const getStoryItem = useCallback(async () => {
 		try {
 			const response = await getStory(id);
 			const result = response.data;
@@ -52,18 +50,49 @@ export default function StoryDetailPage() {
 		} catch (err) {
 			navigate("/404");
 		}
-	};
+	}, [id, navigate, setGenre, setVoice])
+
+	useEffect(() => {
+		getStoryItem();
+	}, [getStoryItem]);
+
 	// 번역
 	const transLang = () => {
 		setLang((prev) => !prev);
 	};
 	// 이야기 삭제
-	const clickDelete = async () => {
-		const response = await deleteStory(id);
-		if (response.status === 200) {
-			alert("삭제가 완료되었습니다.");
-			navigate("/library");
-		}
+	const clickDelete = () => {
+		Swal.fire({
+			title: 'Are you sure?',
+			text: "You won't be able to revert this!",
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Yes, delete it!'
+		}).then(async (result) => {
+			console.log(result);
+			if (result.isConfirmed) {
+				try {
+					const response = await deleteStory(id);
+					if (response.status === 200) {
+						Swal.fire(
+							'Deleted!',
+							'Your file has been deleted.',
+							'success'
+						)
+						navigate("/library");
+					}
+				} catch (error) {
+					console.log(error)
+					Swal.fire({
+						icon: 'error',
+						title: 'Oops...',
+						text: '이야기 삭제를 실패했습니다',
+					})
+				}
+			}
+		})
 	};
 
 	return (
